Fix resetting service start time on session change.
We're expecting a Unix timestamp here, not a JS one.

In general, timestamps are a bit of a mess - the original Timing71 was
written in Python and used Unix timestamps exclusively (so all prior
recordings also use Unix timestamps). Some parts of the JS rewrite use
JS timestamps and others use Unix timestamps. Some sort of rationality
would be nice...